+++ lib/src/selector/list_wildcard.dart
@@ -1,10 +1,10 @@
-import 'package:json_path/src/result.dart';
+import 'package:json_path/src/json_path_match.dart';
 import 'package:json_path/src/selector/selector.dart';
 import 'package:json_path/src/selector/selector_mixin.dart';
 
 class ListWildcard with SelectorMixin implements Selector {
   @override
-  Iterable<Result> filter(Iterable<Result> results) => results
+  Iterable<JsonPathMatch> read(Iterable<JsonPathMatch> matches) => matches
       .where((r) => r.value is List)
       .map((r) => _wrap(r.value, r.path))
       .expand((_) => _);
@@ -12,17 +12,13 @@ class ListWildcard with SelectorMixin implements Selector {
   @override
   String expression() => '[*]';
 
-  Iterable<Result> _wrap(List val, String path) sync* {
+  @override
+  dynamic replace(dynamic json, Replacement replacement) =>
+      (json is List) ? json.map(replacement).toList() : json;
+
+  Iterable<JsonPathMatch> _wrap(List val, String path) sync* {
     for (var i = 0; i < val.length; i++) {
-      yield Result(val[i], path + '[$i]');
+      yield JsonPathMatch(val[i], path + '[$i]');
     }
   }
-
-  @override
-  Object apply(Object json, Object Function(Object _) setter) {
-    // TODO: implement setIn
-    throw UnimplementedError();
-  }
-
-
 }