+++ lib/src/selector/object_union.dart
@@ -1,5 +1,5 @@
-import 'package:json_path/src/quote.dart';
-import 'package:json_path/src/result.dart';
+import 'package:json_path/src/json_path_match.dart';
+import 'package:json_path/src/selector/quote.dart';
 import 'package:json_path/src/selector/selector.dart';
 import 'package:json_path/src/selector/selector_mixin.dart';
 
@@ -9,28 +9,32 @@ class ObjectUnion with SelectorMixin implements Selector {
   final List<String> keys;
 
   @override
-  Iterable<Result> filter(Iterable<Result> results) => results
-      .map((r) => (r.value is Map) ? _map(r.value, r.path) : <Result>[])
+  Iterable<JsonPathMatch> read(Iterable<JsonPathMatch> matches) => matches
+      .map((r) =>
+          (r.value is Map) ? _readMap(r.value, r.path) : <JsonPathMatch>[])
       .expand((_) => _);
 
   @override
   String expression() => '[${keys.map((k) => Quote(k)).join(',')}]';
 
-  Iterable<Result> _map(Map map, String path) => keys
-      .where(map.containsKey)
-      .map((key) => Result(map[key], path + '[${Quote(key)}]'));
-
   @override
-  dynamic apply(dynamic json, dynamic Function(dynamic _) mutate) {
+  dynamic replace(dynamic json, Replacement replacement) {
     if (json is Map) {
-      final diff = Map.fromEntries(keys
-          .where(json.containsKey)
-          .map((key) => MapEntry(key, mutate(json[key]))));
-      if (diff.isEmpty) {
+      final patch = _makePatch(json, replacement);
+      if (patch.isEmpty) {
         return json;
       }
-      return {...json, ...diff};
+      return {...json, ...patch};
     }
     return json;
   }
+
+  Iterable<JsonPathMatch> _readMap(Map map, String path) => keys
+      .where(map.containsKey)
+      .map((key) => JsonPathMatch(map[key], path + '[${Quote(key)}]'));
+
+  Map<String, dynamic> _makePatch(Map map, Replacement replacement) =>
+      Map.fromEntries(keys
+          .where(map.containsKey)
+          .map((key) => MapEntry(key, replacement(map[key]))));
 }