+++ lib/src/selector/root.dart
@@ -1,4 +1,4 @@
-import 'package:json_path/src/result.dart';
+import 'package:json_path/src/json_path_match.dart';
 import 'package:json_path/src/selector/selector.dart';
 import 'package:json_path/src/selector/selector_mixin.dart';
 
@@ -6,14 +6,12 @@ class RootSelector with SelectorMixin implements Selector {
   const RootSelector();
 
   @override
-  Iterable<Result> filter(Iterable<Result> results) =>
-      results.map((m) => Result(m.value, expression()));
+  Iterable<JsonPathMatch> read(Iterable<JsonPathMatch> matches) =>
+      matches.map((m) => JsonPathMatch(m.value, expression()));
 
   @override
   String expression() => r'$';
 
   @override
-  dynamic apply(dynamic json, Function(dynamic _) mutate) {
-    return mutate(json);
-  }
+  dynamic replace(dynamic json, Replacement replacement) => replacement(json);
 }