+++ test/filter_test.dart
@@ -10,38 +10,38 @@ void main() {
     test('Only root', () {
       final root = JsonPath(r'$');
       expect(root.toString(), r'$');
-      expect(root.filter(json).single.value, json);
-      expect(root.filter(json).single.path, r'$');
+      expect(root.read(json).single.value, json);
+      expect(root.read(json).single.path, r'$');
     });
 
     test('Single field', () {
       final store = JsonPath(r'$.store');
       expect(store.toString(), r"$['store']");
-      expect(store.filter(json).single.value, json['store']);
-      expect(store.filter(json).single.path, r"$['store']");
+      expect(store.read(json).single.value, json['store']);
+      expect(store.read(json).single.path, r"$['store']");
     });
 
     test('Single field with digits', () {
       final j = {'aA_12': 'foo'};
       final store = JsonPath(r'$aA_12');
       expect(store.toString(), r"$['aA_12']");
-      expect(store.filter(j).single.value, 'foo');
-      expect(store.filter(j).single.path, r"$['aA_12']");
+      expect(store.read(j).single.value, 'foo');
+      expect(store.read(j).single.path, r"$['aA_12']");
     });
 
     test('Single field in bracket notation', () {
       final store = JsonPath(r"$['store']");
       expect(store.toString(), r"$['store']");
-      expect(store.filter(json).single.value, json['store']);
-      expect(store.filter(json).single.path, r"$['store']");
+      expect(store.read(json).single.value, json['store']);
+      expect(store.read(json).single.path, r"$['store']");
     });
 
     test('Mixed brackets and fields', () {
       final price = JsonPath(r"$['store'].bicycle['price']");
       expect(price.toString(), r"$['store']['bicycle']['price']");
       expect(
-          price.filter(json).single.value, json['store']['bicycle']['price']);
-      expect(price.filter(json).single.path, r"$['store']['bicycle']['price']");
+          price.read(json).single.value, json['store']['bicycle']['price']);
+      expect(price.read(json).single.path, r"$['store']['bicycle']['price']");
     });
   });
 
@@ -56,113 +56,113 @@ void main() {
     test('1:3', () {
       final slice = JsonPath(r'$[1:3]');
       expect(slice.toString(), r'$[1:3]');
-      expect(slice.filter(abc).length, 2);
-      expect(slice.filter(abc).first.value, 'b');
-      expect(slice.filter(abc).first.path, r'$[1]');
-      expect(slice.filter(abc).last.value, 'c');
-      expect(slice.filter(abc).last.path, r'$[2]');
+      expect(slice.read(abc).length, 2);
+      expect(slice.read(abc).first.value, 'b');
+      expect(slice.read(abc).first.path, r'$[1]');
+      expect(slice.read(abc).last.value, 'c');
+      expect(slice.read(abc).last.path, r'$[2]');
     });
     test('1:5:2', () {
       final slice = JsonPath(r'$[1:5:2]');
       expect(slice.toString(), r'$[1:5:2]');
-      expect(slice.filter(abc).length, 2);
-      expect(slice.filter(abc).first.value, 'b');
-      expect(slice.filter(abc).first.path, r'$[1]');
-      expect(slice.filter(abc).last.value, 'd');
-      expect(slice.filter(abc).last.path, r'$[3]');
+      expect(slice.read(abc).length, 2);
+      expect(slice.read(abc).first.value, 'b');
+      expect(slice.read(abc).first.path, r'$[1]');
+      expect(slice.read(abc).last.value, 'd');
+      expect(slice.read(abc).last.path, r'$[3]');
     });
     test('1:5:-2', () {
       final slice = JsonPath(r'$[1:5:-2]');
       expect(slice.toString(), r'$[1:5:-2]');
-      expect(slice.filter(abc).length, 0);
+      expect(slice.read(abc).length, 0);
     });
     test(':3', () {
       final slice = JsonPath(r'$[:3]');
       expect(slice.toString(), r'$[:3]');
-      expect(slice.filter(abc).length, 3);
-      expect(slice.filter(abc).first.value, 'a');
-      expect(slice.filter(abc).first.path, r'$[0]');
-      expect(slice.filter(abc).last.value, 'c');
-      expect(slice.filter(abc).last.path, r'$[2]');
+      expect(slice.read(abc).length, 3);
+      expect(slice.read(abc).first.value, 'a');
+      expect(slice.read(abc).first.path, r'$[0]');
+      expect(slice.read(abc).last.value, 'c');
+      expect(slice.read(abc).last.path, r'$[2]');
     });
     test(':3:2', () {
       final slice = JsonPath(r'$[:3:2]');
       expect(slice.toString(), r'$[:3:2]');
-      expect(slice.filter(abc).length, 2);
-      expect(slice.filter(abc).first.value, 'a');
-      expect(slice.filter(abc).first.path, r'$[0]');
-      expect(slice.filter(abc).last.value, 'c');
-      expect(slice.filter(abc).last.path, r'$[2]');
+      expect(slice.read(abc).length, 2);
+      expect(slice.read(abc).first.value, 'a');
+      expect(slice.read(abc).first.path, r'$[0]');
+      expect(slice.read(abc).last.value, 'c');
+      expect(slice.read(abc).last.path, r'$[2]');
     });
     test('3::2', () {
       final slice = JsonPath(r'$[3::2]');
       expect(slice.toString(), r'$[3::2]');
-      expect(slice.filter(abc).length, 2);
-      expect(slice.filter(abc).first.value, 'd');
-      expect(slice.filter(abc).first.path, r'$[3]');
-      expect(slice.filter(abc).last.value, 'f');
-      expect(slice.filter(abc).last.path, r'$[5]');
+      expect(slice.read(abc).length, 2);
+      expect(slice.read(abc).first.value, 'd');
+      expect(slice.read(abc).first.path, r'$[3]');
+      expect(slice.read(abc).last.value, 'f');
+      expect(slice.read(abc).last.path, r'$[5]');
     });
     test('100:', () {
       final slice = JsonPath(r'$[100:]');
       expect(slice.toString(), r'$[100:]');
-      expect(slice.filter(abc).length, 0);
+      expect(slice.read(abc).length, 0);
     });
     test('3:', () {
       final slice = JsonPath(r'$[3:]');
       expect(slice.toString(), r'$[3:]');
-      expect(slice.filter(abc).length, 4);
-      expect(slice.filter(abc).first.value, 'd');
-      expect(slice.filter(abc).first.path, r'$[3]');
-      expect(slice.filter(abc).last.value, 'g');
-      expect(slice.filter(abc).last.path, r'$[6]');
+      expect(slice.read(abc).length, 4);
+      expect(slice.read(abc).first.value, 'd');
+      expect(slice.read(abc).first.path, r'$[3]');
+      expect(slice.read(abc).last.value, 'g');
+      expect(slice.read(abc).last.path, r'$[6]');
     });
     test(':-5', () {
       final slice = JsonPath(r'$[:-5]');
       expect(slice.toString(), r'$[:-5]');
-      expect(slice.filter(abc).length, 2);
-      expect(slice.filter(abc).first.value, 'a');
-      expect(slice.filter(abc).first.path, r'$[0]');
-      expect(slice.filter(abc).last.value, 'b');
-      expect(slice.filter(abc).last.path, r'$[1]');
+      expect(slice.read(abc).length, 2);
+      expect(slice.read(abc).first.value, 'a');
+      expect(slice.read(abc).first.path, r'$[0]');
+      expect(slice.read(abc).last.value, 'b');
+      expect(slice.read(abc).last.path, r'$[1]');
     });
 
     test('-5:', () {
       final slice = JsonPath(r'$[-5:]');
       expect(slice.toString(), r'$[-5:]');
-      expect(slice.filter(abc).length, 5);
-      expect(slice.filter(abc).first.value, 'c');
-      expect(slice.filter(abc).first.path, r'$[2]');
-      expect(slice.filter(abc).last.value, 'g');
-      expect(slice.filter(abc).last.path, r'$[6]');
+      expect(slice.read(abc).length, 5);
+      expect(slice.read(abc).first.value, 'c');
+      expect(slice.read(abc).first.path, r'$[2]');
+      expect(slice.read(abc).last.value, 'g');
+      expect(slice.read(abc).last.path, r'$[6]');
     });
     test('0:6', () {
       final slice = JsonPath(r'$[0:6]');
       expect(slice.toString(), r'$[:6]');
-      expect(slice.filter(abc).length, 6);
-      expect(slice.filter(abc).first.value, 'a');
-      expect(slice.filter(abc).first.path, r'$[0]');
-      expect(slice.filter(abc).last.value, 'f');
-      expect(slice.filter(abc).last.path, r'$[5]');
+      expect(slice.read(abc).length, 6);
+      expect(slice.read(abc).first.value, 'a');
+      expect(slice.read(abc).first.path, r'$[0]');
+      expect(slice.read(abc).last.value, 'f');
+      expect(slice.read(abc).last.path, r'$[5]');
     });
     test('0:100', () {
       final slice = JsonPath(r'$[0:100]');
       expect(slice.toString(), r'$[:100]');
-      expect(slice.filter(abc).length, 7);
-      expect(slice.filter(abc).first.value, 'a');
-      expect(slice.filter(abc).first.path, r'$[0]');
-      expect(slice.filter(abc).last.value, 'g');
-      expect(slice.filter(abc).last.path, r'$[6]');
+      expect(slice.read(abc).length, 7);
+      expect(slice.read(abc).first.value, 'a');
+      expect(slice.read(abc).first.path, r'$[0]');
+      expect(slice.read(abc).last.value, 'g');
+      expect(slice.read(abc).last.path, r'$[6]');
     });
 
     test('-6:-1', () {
       final slice = JsonPath(r'$[-6:-1]');
       expect(slice.toString(), r'$[-6:-1]');
-      expect(slice.filter(abc).length, 5);
-      expect(slice.filter(abc).first.value, 'b');
-      expect(slice.filter(abc).first.path, r'$[1]');
-      expect(slice.filter(abc).last.value, 'f');
-      expect(slice.filter(abc).last.path, r'$[5]');
+      expect(slice.read(abc).length, 5);
+      expect(slice.read(abc).first.value, 'b');
+      expect(slice.read(abc).first.path, r'$[1]');
+      expect(slice.read(abc).last.value, 'f');
+      expect(slice.read(abc).last.path, r'$[5]');
     });
   });
 
@@ -171,7 +171,7 @@ void main() {
       final j = {r"sq'sq s\s qs\'qs": 'value'};
       final path = JsonPath(r"$['sq\'sq s\\s qs\\\'qs']");
       expect(path.toString(), r"$['sq\'sq s\\s qs\\\'qs']");
-      final select = path.filter(j);
+      final select = path.read(j);
       expect(select.single.value, 'value');
       expect(select.single.path, r"$['sq\'sq s\\s qs\\\'qs']");
     });
@@ -182,11 +182,11 @@ void main() {
       final abc = 'abcdefg'.split('');
       final union = JsonPath(r'$[2,3,100,5]');
       expect(union.toString(), r'$[2,3,100,5]');
-      expect(union.filter(abc).length, 3);
-      expect(union.filter(abc).first.value, 'c');
-      expect(union.filter(abc).first.path, r'$[2]');
-      expect(union.filter(abc).last.value, 'f');
-      expect(union.filter(abc).last.path, r'$[5]');
+      expect(union.read(abc).length, 3);
+      expect(union.read(abc).first.value, 'c');
+      expect(union.read(abc).first.path, r'$[2]');
+      expect(union.read(abc).last.value, 'f');
+      expect(union.read(abc).last.path, r'$[5]');
     });
     test('Object', () {
       final abc = {
@@ -196,11 +196,11 @@ void main() {
       };
       final union = JsonPath(r"$['a','x',c]");
       expect(union.toString(), r"$['a','x','c']");
-      expect(union.filter(abc).length, 2);
-      expect(union.filter(abc).first.value, 'A');
-      expect(union.filter(abc).first.path, r"$['a']");
-      expect(union.filter(abc).last.value, 'C');
-      expect(union.filter(abc).last.path, r"$['c']");
+      expect(union.read(abc).length, 2);
+      expect(union.read(abc).first.value, 'A');
+      expect(union.read(abc).first.path, r"$['a']");
+      expect(union.read(abc).last.value, 'C');
+      expect(union.read(abc).last.path, r"$['c']");
     });
   });
 
@@ -208,19 +208,19 @@ void main() {
     test('All in root', () {
       final allInRoot = JsonPath(r'$.*');
       expect(allInRoot.toString(), r'$.*');
-      expect(allInRoot.filter(json).length, 1);
-      expect(allInRoot.filter(json).single.value, json['store']);
-      expect(allInRoot.filter(json).single.path, r"$['store']");
+      expect(allInRoot.read(json).length, 1);
+      expect(allInRoot.read(json).single.value, json['store']);
+      expect(allInRoot.read(json).single.path, r"$['store']");
     });
 
     test('All in store', () {
       final allInStore = JsonPath(r'$.store.*');
       expect(allInStore.toString(), r"$['store'].*");
-      expect(allInStore.filter(json).length, 2);
-      expect(allInStore.filter(json).first.value, json['store']['book']);
-      expect(allInStore.filter(json).first.path, r"$['store']['book']");
-      expect(allInStore.filter(json).last.value, json['store']['bicycle']);
-      expect(allInStore.filter(json).last.path, r"$['store']['bicycle']");
+      expect(allInStore.read(json).length, 2);
+      expect(allInStore.read(json).first.value, json['store']['book']);
+      expect(allInStore.read(json).first.path, r"$['store']['book']");
+      expect(allInStore.read(json).last.value, json['store']['bicycle']);
+      expect(allInStore.read(json).last.path, r"$['store']['bicycle']");
     });
   });
 
@@ -228,31 +228,31 @@ void main() {
     test('Recursive', () {
       final allNode = JsonPath(r'$..');
       expect(allNode.toString(), r'$..');
-      expect(allNode.filter(json).length, 8);
-      expect(allNode.filter(json).first.value, json);
-      expect(allNode.filter(json).first.path, r'$');
-      expect(allNode.filter(json).last.value, json['store']['bicycle']);
-      expect(allNode.filter(json).last.path, r"$['store']['bicycle']");
+      expect(allNode.read(json).length, 8);
+      expect(allNode.read(json).first.value, json);
+      expect(allNode.read(json).first.path, r'$');
+      expect(allNode.read(json).last.value, json['store']['bicycle']);
+      expect(allNode.read(json).last.path, r"$['store']['bicycle']");
     });
 
     test('Recursive with all values', () {
       final path = JsonPath(r'$..*');
       expect(path.toString(), r'$..*');
-      expect(path.filter(json).length, 27);
-      expect(path.filter(json).first.value, json['store']);
-      expect(path.filter(json).first.path, r"$['store']");
-      expect(path.filter(json).last.value, json['store']['bicycle']['price']);
-      expect(path.filter(json).last.path, r"$['store']['bicycle']['price']");
+      expect(path.read(json).length, 27);
+      expect(path.read(json).first.value, json['store']);
+      expect(path.read(json).first.path, r"$['store']");
+      expect(path.read(json).last.value, json['store']['bicycle']['price']);
+      expect(path.read(json).last.path, r"$['store']['bicycle']['price']");
     });
 
     test('Every price tag', () {
       final path = JsonPath(r'$..price');
       expect(path.toString(), r"$..['price']");
-      expect(path.filter(json).length, 5);
-      expect(path.filter(json).first.value, json['store']['book'][0]['price']);
-      expect(path.filter(json).first.path, r"$['store']['book'][0]['price']");
-      expect(path.filter(json).last.value, json['store']['bicycle']['price']);
-      expect(path.filter(json).last.path, r"$['store']['bicycle']['price']");
+      expect(path.read(json).length, 5);
+      expect(path.read(json).first.value, json['store']['book'][0]['price']);
+      expect(path.read(json).first.path, r"$['store']['book'][0]['price']");
+      expect(path.read(json).last.value, json['store']['bicycle']['price']);
+      expect(path.read(json).last.path, r"$['store']['bicycle']['price']");
     });
   });
 
@@ -260,25 +260,25 @@ void main() {
     test('Path with an index', () {
       final path = JsonPath(r'$.store.book[0].title');
       expect(path.toString(), r"$['store']['book'][0]['title']");
-      expect(path.filter(json).single.value, 'Sayings of the Century');
-      expect(path.filter(json).single.path, r"$['store']['book'][0]['title']");
+      expect(path.read(json).single.value, 'Sayings of the Century');
+      expect(path.read(json).single.path, r"$['store']['book'][0]['title']");
     });
 
     test('Last element of array (regression #1)', () {
       final path = JsonPath(r"$['store']['book'][3]['price']");
       expect(path.toString(), r"$['store']['book'][3]['price']");
-      expect(path.filter(json).single.value, 22.99);
-      expect(path.filter(json).single.path, r"$['store']['book'][3]['price']");
+      expect(path.read(json).single.value, 22.99);
+      expect(path.read(json).single.path, r"$['store']['book'][3]['price']");
     });
 
     test('All in list', () {
       final path = JsonPath(r'$.store.book[*]');
       expect(path.toString(), r"$['store']['book'][*]");
-      expect(path.filter(json).length, 4);
-      expect(path.filter(json).first.value, json['store']['book'][0]);
-      expect(path.filter(json).first.path, r"$['store']['book'][0]");
-      expect(path.filter(json).last.value, json['store']['book'][3]);
-      expect(path.filter(json).last.path, r"$['store']['book'][3]");
+      expect(path.read(json).length, 4);
+      expect(path.read(json).first.value, json['store']['book'][0]);
+      expect(path.read(json).first.path, r"$['store']['book'][0]");
+      expect(path.read(json).last.value, json['store']['book'][3]);
+      expect(path.read(json).last.path, r"$['store']['book'][3]");
     });
   });
 
@@ -288,27 +288,36 @@ void main() {
         'discounted': (e) => e is Map && e['price'] is num && e['price'] < 20
       });
       expect(path.toString(), r"$['store']..[?discounted]");
-      expect(path.filter(json).length, 4);
-      expect(path.filter(json).first.value, json['store']['book'][0]);
-      expect(path.filter(json).first.path, r"$['store']['book'][0]");
-      expect(path.filter(json).last.value, json['store']['bicycle']);
-      expect(path.filter(json).last.path, r"$['store']['bicycle']");
+      expect(path.read(json).length, 4);
+      expect(path.read(json).first.value, json['store']['book'][0]);
+      expect(path.read(json).first.path, r"$['store']['book'][0]");
+      expect(path.read(json).last.value, json['store']['bicycle']);
+      expect(path.read(json).last.path, r"$['store']['bicycle']");
     });
 
     test('Can be applied to scalars', () {
-      final path = JsonPath(r'$.store..price[?low]', filter: {
-        'low': (e) => e is num && e < 20
-      });
+      final path = JsonPath(r'$.store..price[?low]',
+          filter: {'low': (e) => e is num && e < 20});
       expect(path.toString(), r"$['store']..['price'][?low]");
-      expect(path.filter(json).length, 4);
-      expect(path.filter(json).first.value, json['store']['book'][0]['price']);
-      expect(path.filter(json).first.path, r"$['store']['book'][0]['price']");
-      expect(path.filter(json).last.value, json['store']['bicycle']['price']);
-      expect(path.filter(json).last.path, r"$['store']['bicycle']['price']");
+      expect(path.read(json).length, 4);
+      expect(path.read(json).first.value, json['store']['book'][0]['price']);
+      expect(path.read(json).first.path, r"$['store']['book'][0]['price']");
+      expect(path.read(json).last.value, json['store']['bicycle']['price']);
+      expect(path.read(json).last.path, r"$['store']['bicycle']['price']");
     });
 
     test('Missing filter', () {
       expect(() => JsonPath(r'$.store..[?discounted]'), throwsFormatException);
     });
+  });
+  test('Modifying in-place', () {
+    final someBooks = JsonPath(r'$.store.book[::2]');
+    someBooks.read(json).forEach((result) {
+      result.value['title'] = 'Banana';
+    });
+    expect(json['store']['book'][0]['title'], 'Banana');
+    expect(json['store']['book'][1]['title'], 'Sword of Honour');
+    expect(json['store']['book'][2]['title'], 'Banana');
+    expect(json['store']['book'][3]['title'], 'The Lord of the Rings');
   });
 }